Added score range metrics for genesys formats

+++ SKCSwift/Models/YGO.swift
@@ -152,7 +152,7 @@ nonisolated struct BanListDate: Codable, Hashable {
 
 nonisolated struct BannedContent: Codable, Equatable {
     let forbidden, limited, semiLimited: [YGOCard]
-    let numForbidden, numLimited, numSemiLimited, numLimitedOne, numLimitedTwo, numLimitedThree: UInt8
+    let numForbidden, numLimited, numSemiLimited, numLimitedOne, numLimitedTwo, numLimitedThree: UInt16
 }
 
 /*
